+++ webapp/src/main/resources/public/js/utils/JobStatus.js
@@ -1,4 +1,3 @@
-
 export const JobStatus = Object.freeze({
     IN_PROGRESS: "IN_PROGRESS",
     SUCCEEDED: "SUCCEEDED",
